+++ PortableHoles/CHANGELOG.md
@@ -1,5 +1,16 @@
 # Portable Holes Change Log
 
+## 1.2.0 (September 17, 2022)
+
+### Fixed
+
+* Crafting recipe drawn as a hole instead of stairs.
+* Crafted object will show the correct name and description.
+
+### Changed
+
+* New texture thanks to [Tai](https://www.nexusmods.com/stardewvalley/users/92060238).
+
 ## 1.1.0 (August 24, 2022)
 
 ### Added